requirements.txt
flask
python-docx
flask-cors
fastapi
requests
uvicorn
python-dotenv